% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AgroTech-package}
\alias{AgroTech}
\alias{AgroTech-package}
\title{AgroTech: Data Analysis of Pesticide Application Technology}
\description{
In total it has 7 functions, three for calculating machine calibration, which determine application rate (L/ha), nozzle flow (L/min) and amount of product (L or kg) to be added. to the tank with each sprayer filling. Two functions for graphs of the flow distribution of the nozzles (L/min) in the application bar and, of the temporal variability of the meteorological conditions (air temperature, relative humidity of the air and wind speed). Two functions to determine the spray deposit (uL/cm2), through the methodology called spectrophotometry, with the aid of bright blue (Palladini, L.A., Raetano, C.G., Velini, E.D. (2005), \doi{10.1590/S0103-90162005000500005}) or metallic markers (Chaim, A., Castro, V.L.S.S., Correles, F.M., Galvão, J.A.H., Cabral, O.M.R., Nicolella, G. (1999), \doi{10.1590/S0100-204X1999000500003}). The package supports the analysis and representation of information, using a single free software that meets the most diverse areas of activity in application technology.
}
\author{
\strong{Maintainer}: Rodrigo Yudi Palhaci Marubayashi \email{marubayashi@uel.br} (\href{https://orcid.org/0000-0003-2778-8654}{ORCID})

Authors:
\itemize{
  \item Gabriel Danilo Shimizu (\href{https://orcid.org/0000-0001-8524-508X}{ORCID}) [contributor]
  \item Otavio Jorge Grigoli Abi Saab (\href{https://orcid.org/0000-0002-1757-636X}{ORCID}) [contributor]
}

}
\keyword{internal}
