% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_markblue.R
\docType{data}
\name{example_markblue}
\alias{example_markblue}
\title{Dataset: Example markblue}
\format{
data.frame containing data set
  \describe{
  \item{\code{curva}}{Vector with curves}
  \item{\code{TRATAMENTO}}{Numeric vector with treatment}
  \item{\code{repe}}{Numeric vector with repetition}
  \item{\code{Abs}}{Numeric vector with absorbance}
  \item{\code{area}}{Numeric vector with area}
  }
}
\usage{
data(example_markblue)
}
\description{
The data are part of an experiment that studied the spray deposit
}
\examples{
data(example_markblue)
}
\author{
Rodrigo Yudi Palhaci Marubayashi, \email{marubayashi@uel.br}

Gabriel Danilo Shimizu

Otavio Jorge Grigoli Abi Saab
}
\keyword{datasets}
