% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_markbluecurve.R
\docType{data}
\name{example_markbluecurve}
\alias{example_markbluecurve}
\title{Dataset: Example markbluecurve}
\format{
data.frame containing data set
  \describe{
  \item{\code{curva}}{Vector with curve}
  \item{\code{Amostra}}{Numeric vector with sample}
  \item{\code{Abs}}{Numeric vector with absorbance}
  \item{\code{ppm}}{Numeric vector with concentration}
  }
}
\usage{
data(example_markbluecurve)
}
\description{
The data are part of an experiment that studied the spray deposit
}
\examples{
data(example_markbluecurve)
}
\author{
Rodrigo Yudi Palhaci Marubayashi, \email{marubayashi@uel.br}

Gabriel Danilo Shimizu

Otavio Jorge Grigoli Abi Saab
}
\keyword{datasets}
