% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_markmet.R
\docType{data}
\name{example_markmet}
\alias{example_markmet}
\title{Dataset: Example markmet}
\format{
data.frame containing data set
  \describe{
  \item{\code{trat}}{Vector with treatment}
  \item{\code{repe}}{Numeric vector with repetition}
  \item{\code{ppm}}{Numeric vector with concentration}
  }
}
\usage{
data(example_markmet)
}
\description{
The data are part of an experiment that studied the spray deposit
}
\examples{
data(example_markmet)
}
\author{
Rodrigo Yudi Palhaci Marubayashi, \email{marubayashi@uel.br}

Gabriel Danilo Shimizu

Otavio Jorge Grigoli Abi Saab
}
\keyword{datasets}
