% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_meteorological.R
\docType{data}
\name{example_meteorological}
\alias{example_meteorological}
\title{Dataset: Example meteorological}
\format{
data.frame containing data set
  \describe{
  \item{\code{tempo}}{Numeric vector with time}
  \item{\code{temp}}{Numeric vector with air temperature}
  \item{\code{ur}}{Numeric vector with relative humidity}
  \item{\code{vento}}{Numeric vector with wind speed}
  }
}
\usage{
data("example_meteorological")
}
\description{
The data come from a meteorological station on a rural property in the city of Rolandia/PR
}
\examples{
data(example_meteorological)
}
\author{
Rodrigo Yudi Palhaci Marubayashi, \email{marubayashi@uel.br}

Gabriel Danilo Shimizu

Otavio Jorge Grigoli Abi Saab
}
\keyword{datasets}
