% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartemp.R
\name{vartemp}
\alias{vartemp}
\title{Temporal variability graph of weather conditions}
\usage{
vartemp(
  file,
  nx = 1,
  ny = 2,
  variable = NA,
  ylab = "Dependent",
  xlab = "Independent",
  size.text = 12,
  size.title = 12,
  size.strip = 12,
  size.lty = 0.7
)
}
\arguments{
\item{file}{Excel file (xlsx)}

\item{nx}{Time}

\item{ny}{Weather conditions}

\item{variable}{Variable name}

\item{ylab}{y axis (Dependent)}

\item{xlab}{x axis (Independent)}

\item{size.text}{Size text (\emph{default} is 12)}

\item{size.title}{Size title (\emph{default} is 12)}

\item{size.strip}{Size strip (\emph{default} is 12)}

\item{size.lty}{Size line (\emph{default} is 0.7)}
}
\value{
Returns graph of ggplot2
}
\description{
This is a function to check weather conditions in agricultural spraying
}
\examples{
data("example_meteorological")
vartemp(example_meteorological)
}
\references{
No reference
}
\author{
Rodrigo Yudi Palhaci Marubayashi, \email{marubayashi@uel.br}

Gabriel Danilo Shimizu

Otavio Jorge Grigoli Abi Saab
}
